// src/app/models/user-interface.ts
export interface UserInterface {
  id?: number;
  name?: string;
  last_name?: string;
  dni?: string;
  birthday?: Date;
  mobile?: number;
  address?: string;
  email?: string;
  username?: string;
  avatar?: string;
  role_id?: number;
  branch_id?: number;
  status?: number;
  password?: string;
  password_confirmation?: string;
}